Verilog: split synthesis into behavioral and structural
Split verilog_synthesis.cpp into three implementation files:
- verilog_synthesis.cpp: expressions, module-item dispatch, assertions,
  continuous assigns, declarations, top-level driver
- verilog_synthesis_behavioral.cpp: statement synthesis, control flow,
  assignments, function inlining
- verilog_synthesis_structural.cpp: module instances, port connections,
  built-in gates

Reorganize verilog_synthesis_class.h with clear section separation
for behavioral state, behavioral methods, structural methods, and
module-item synthesis. Add verilog_synthesis_behavioral.h defining
the behavioral state types as standalone classes.
