+++ src/verilog/verilog_synthesis.cpp
@@ -312,220 +312,6 @@ exprt verilog_synthesist::value_mapt::guarded_expr(exprt expr) const
 
 /*******************************************************************\
 
-Function: verilog_synthesist::function_locality
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::function_locality(const symbolt &function_symbol)
-{
-  // Find all local symbols of the function, mark their
-  // assignments as local.
-  auto prefix = id2string(function_symbol.name) + '.';
-  for(auto &symbol : symbol_table.symbols)
-  {
-    if(symbol.first.starts_with(prefix))
-    {
-      assignments[symbol.first].is_cycle_local = true;
-    }
-  }
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::expand_function_call
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-exprt verilog_synthesist::expand_function_call(
-  const function_call_exprt &call,
-  symbol_statet symbol_state)
-{
-  // Is it a 'system function call'?
-  if(call.is_system_function_call())
-  {
-    auto base_name = to_verilog_identifier_expr(call.function()).base_name();
-    if(base_name == "$ND")
-    {
-      std::string identifier =
-        id2string(module) + "::nondet::" + std::to_string(nondet_count++);
-
-      auto arguments = call.arguments();
-      exprt select_one(
-        ID_constraint_select_one, call.type(), std::move(arguments));
-      select_one.set(ID_identifier, identifier);
-      return select_one.with_source_location(call);
-    }
-    else if(base_name == "$past")
-    {
-      auto what = synth_expr_rec(call.arguments()[0], symbol_state);
-      auto ticks = call.arguments().size() < 2
-                     ? from_integer(1, integer_typet())
-                     : call.arguments()[1];
-      return verilog_past_exprt{what, ticks}.with_source_location(call);
-    }
-    else if(
-      base_name == "$stable" || base_name == "$rose" || base_name == "$fell" ||
-      base_name == "$changed")
-    {
-      DATA_INVARIANT(call.arguments().size() >= 1, "must have argument");
-      auto what = synth_expr_rec(call.arguments()[0], symbol_state);
-      auto past = verilog_past_exprt{what, from_integer(1, integer_typet())}
-                    .with_source_location(call);
-
-      auto lsb = [](exprt expr) {
-        return extractbit_exprt{
-          std::move(expr), from_integer(0, integer_typet{})};
-      };
-
-      if(base_name == "$stable")
-        return equal_exprt{what, past};
-      else if(base_name == "$changed")
-        return notequal_exprt{what, past};
-      else if(base_name == "$rose")
-        return and_exprt{not_exprt{lsb(past)}, lsb(what)};
-      else if(base_name == "$fell")
-        return and_exprt{lsb(past), not_exprt{lsb(what)}};
-      else
-        DATA_INVARIANT(false, "all cases covered");
-    }
-    else if(base_name == "$countones")
-    {
-      // lower to popcount
-      DATA_INVARIANT(
-        call.arguments().size() == 1, "$countones must have one argument");
-      auto what = synth_expr(call.arguments()[0], symbol_state); // rec. call
-      return popcount_exprt{what, call.type()};
-    }
-    else
-    {
-      // Attempt to constant fold.
-      verilog_typecheck_exprt verilog_typecheck_expr(
-        standard, false, ns, get_message_handler());
-      auto result =
-        verilog_typecheck_expr.elaborate_constant_system_function_call(call);
-      if(!result.is_constant())
-      {
-        throw errort().with_location(call.source_location())
-          << "cannot synthesise system function " << to_string(call.function());
-      }
-
-      return result;
-    }
-  }
-
-  // check some restrictions
-  if(construct == constructt::OTHER)
-  {
-    throw errort().with_location(call.source_location())
-      << "function call not allowed here";
-  }
-
-  // this is essentially inlined
-  const symbol_exprt &function=to_symbol_expr(call.function());
-
-  const symbolt &symbol=ns.lookup(function);
-  
-  if(symbol.type.id()!=ID_code)
-  {
-    throw errort().with_location(call.source_location())
-      << "function call expects function argument";
-  }
-  
-  const code_typet &code_type=to_code_type(symbol.type);
-
-  if(code_type.return_type()==empty_typet())
-  {
-    throw errort().with_location(call.source_location())
-      << "function call cannot call task";
-  }
-
-  // preserve the previous call frame, if any
-  auto old_tf_frame = tf_frame;
-
-  // remember the guard
-  auto entry_guard = value_map->guard;
-
-  // create a fresh call frame
-  tf_frame = tf_framet{};
-
-  const symbolt &return_symbol =
-    ns.lookup(id2string(symbol.name) + "." + id2string(symbol.base_name));
-
-  tf_frame.value().return_value = return_symbol.symbol_expr();
-
-  const code_typet::parameterst &parameters=
-    code_type.parameters();
-    
-  const exprt::operandst &actuals=
-    call.op1().operands();
-    
-  if(parameters.size()!=actuals.size())
-  {
-    throw errort().with_location(call.source_location())
-      << "wrong number of arguments";
-  }
-
-  // preserve locality of function-local variables
-  function_locality(symbol);
-
-  // do assignments to function parameters
-  for(unsigned i=0; i<parameters.size(); i++)
-  {
-    const symbolt &a_symbol=ns.lookup(parameters[i].get_identifier());
-    verilog_blocking_assignt assignment{
-      a_symbol.symbol_expr().with_source_location(call), actuals[i]};
-    assignment.add_source_location()=call.source_location();
-    synth_statement(assignment);
-  }
-
-  for(auto &statement : symbol.value.operands())
-    synth_statement(to_verilog_statement(statement));
-
-  // merge in edges from 'return' statements, if any
-  for(auto &state : tf_frame.value().return_statement_states)
-  {
-    auto guard_expr = conjunction(state.guard);
-    merge(
-      guard_expr, state.current, value_map->current, false, value_map->current);
-    merge(guard_expr, state.final, value_map->final, true, value_map->final);
-  }
-
-  // replace function call by return value symbol
-  auto result = synth_expr(return_symbol.symbol_expr(), symbol_statet::CURRENT);
-
-  // restore the previous task/function frame
-  tf_frame = old_tf_frame;
-  value_map->guard = entry_guard;
-
-  // do assignments to output parameters
-  for(unsigned i = 0; i < parameters.size(); i++)
-  {
-    const symbolt &a_symbol = ns.lookup(parameters[i].get_identifier());
-    if(parameters[i].get_bool(ID_output))
-    {
-      verilog_blocking_assignt assignment{actuals[i], a_symbol.symbol_expr()};
-      assignment.add_source_location() = call.source_location();
-      synth_statement(assignment);
-    }
-  }
-
-  return result;
-}
-
-/*******************************************************************\
-
 Function: verilog_synthesist::expand_hierarchical_identifier
 
   Inputs:
@@ -576,368 +362,6 @@ void verilog_synthesist::expand_hierarchical_identifier(
 
 /*******************************************************************\
 
-Function: verilog_synthesist::assignment_rec
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::assignment_rec(
-  const exprt &lhs, // synth_lhs_expr applied
-  const exprt &rhs, // synth_expr applied
-  bool blocking)
-{
-  if(lhs.id()==ID_concatenation) // split it up                                
-  {
-    // TODO: split it up more intelligently;
-    // bit-wise is wasteful.
-    mp_integer offset = 0;
-
-    // do it from right to left
-    
-    for(exprt::operandst::const_reverse_iterator
-        it=lhs.operands().rbegin();
-        it!=lhs.operands().rend();
-        it++)
-    {
-      auto offset_constant = from_integer(offset, natural_typet{});
-
-      if(it->type().id()==ID_bool)
-      {
-        exprt bit_extract = extractbit_exprt{rhs, offset_constant};
-        ++offset;
-
-        assignment_rec(*it, bit_extract, blocking);
-      }
-      else if(it->type().id()==ID_signedbv ||
-              it->type().id()==ID_unsignedbv)
-      {
-        extractbits_exprt bit_extract(rhs, offset_constant, it->type());
-
-        assignment_rec(*it, bit_extract, blocking);
-
-        auto width = get_width(it->type());
-        offset+=width;
-      }
-      else
-      {
-        throw errort().with_location(it->source_location())
-          << "assignment to type `" << to_string(it->type())
-          << "' not supported";
-      }
-    }
-    
-    return;
-  }
-  
-  // get identifier
-  const symbolt &symbol=assignment_symbol(lhs);
-  
-  if(symbol.type.id()==ID_verilog_realtime)
-  {
-    // we silently ignore these
-    return;
-  }
-
-  if(!symbol.is_lvalue)
-  {
-    throw errort().with_location(lhs.source_location())
-      << "assignment to non-variable";
-  }
-
-  if(construct==constructt::ALWAYS &&
-     event_guard==event_guardt::NONE)
-  {
-    throw errort().with_location(lhs.source_location())
-      << "assignment in 'always' context without event guard";
-  }
-
-  if(construct == constructt::ALWAYS_FF && event_guard == event_guardt::NONE)
-  {
-    throw errort().with_location(lhs.source_location())
-      << "assignment in 'always_ff' context without event guard";
-  }
-
-  {
-    event_guardt new_type;
-
-    if(construct == constructt::INITIAL)
-      new_type=event_guardt::CLOCK;
-    else if(construct == constructt::ALWAYS_COMB)
-      new_type = event_guardt::COMBINATIONAL;
-    else
-      new_type = event_guard;
-
-    assignmentt &assignment=assignments[symbol.name];
-
-    if(assignment.is_cycle_local)
-    {
-    }
-    else
-    {
-      if(assignment.type == event_guardt::NONE)
-        assignment.type = new_type;
-      else if(assignment.type != new_type)
-      {
-        throw errort().with_location(lhs.source_location())
-          << "conflicting assignment types for `" << symbol.base_name
-          << "' (new: " << as_string(new_type)
-          << ", old: " << as_string(assignment.type) << ")";
-      }
-
-      membert member;
-      assignment_member_rec(
-        lhs,
-        member,
-        (construct == constructt::INITIAL) ? assignment.init : assignment.next);
-    }
-  }
-
-  {
-    assert(value_map!=NULL);
-
-    auto new_value = assignment_rec(lhs, rhs); // start of recursion
-
-    if(new_value.is_not_nil())
-    {
-      // These can explode if the symbol is assigned more than once
-      // and also used more than once, e.g., in a multi-dimensional array.
-      // We add a fresh symbol for anything that is not trivial
-      // to prevent that.
-      if(new_value.id() != ID_symbol && new_value.id() != ID_constant)
-      {
-        replace_by_wire(new_value, symbol);
-      }
-
-      if(blocking)
-        value_map->current.assign(symbol.name, new_value);
-
-      value_map->final.assign(symbol.name, new_value);
-    }
-  }
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::assignment_rec
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-exprt verilog_synthesist::assignment_rec(
-  const exprt &lhs, // synth_lhs_expr applied
-  const exprt &rhs) // synth_expr applied
-{
-  if(lhs.id()==ID_symbol)
-  {
-    return rhs;
-  }
-  else if(lhs.id()==ID_index ||
-          lhs.id()==ID_extractbit)
-  {
-    if(lhs.operands().size()!=2)
-    {
-      throw errort() << "index takes two operands";
-    }
-
-    const exprt &lhs_array = to_binary_expr(lhs).op0();
-    const exprt &lhs_index = to_binary_expr(lhs).op1();
-
-    // turn
-    //   a[i]=e
-    // into
-    //   a'==a WITH [i:=e]
-
-    with_exprt new_rhs(lhs_array, lhs_index, rhs);
-
-    // do the array
-    new_rhs.old() = synth_expr(new_rhs.old(), symbol_statet::FINAL);
-
-    return assignment_rec(lhs_array, new_rhs); // recursive call
-  }
-  else if(lhs.id() == ID_verilog_non_indexed_part_select)
-  {
-    // we flatten n-bit part select into n times extractbit
-    auto &part_select = to_verilog_non_indexed_part_select_expr(lhs);
-
-    const exprt &lhs_src = part_select.src();
-    const exprt &lhs_msb = part_select.msb();
-    const exprt &lhs_lsb = part_select.lsb();
-
-    mp_integer from, to;
-
-    if(to_integer_non_constant(lhs_msb, to))
-    {
-      throw errort().with_location(lhs_msb.source_location())
-        << "failed to convert range";
-    }
-
-    if(to_integer_non_constant(lhs_lsb, from))
-    {
-      throw errort().with_location(lhs_lsb.source_location())
-        << "failed to convert range";
-    }
-
-    if(from > to)
-      std::swap(from, to);
-
-    // redundant?
-    if(from == 0 && to == get_width(lhs_src.type()) - 1)
-    {
-      return assignment_rec(lhs_src, rhs); // recursive call
-    }
-
-    // turn
-    //   a[i]=e
-    // into
-    //   a'==a WITH [from:=e[0]] ... WITH [to:=e[to-from]]
-
-    // synthesise 'a'
-    auto new_rhs = synth_expr(lhs_src, symbol_statet::FINAL);
-    const auto rhs_width = get_width(lhs_src.type());
-
-    // We drop bits that are out of bounds
-    auto from_in_range = std::max(mp_integer{0}, from);
-    auto to_in_range = std::min(rhs_width - 1, to);
-
-    // add the WITH expressions for the indexes in the range
-    for(mp_integer i = from_in_range; i <= to_in_range; ++i)
-    {
-      exprt offset = from_integer(i - from, integer_typet{});
-      exprt rhs_extractbit = extractbit_exprt{rhs, std::move(offset)};
-      exprt count = from_integer(i, integer_typet{});
-      new_rhs = with_exprt{
-        std::move(new_rhs), std::move(count), std::move(rhs_extractbit)};
-    }
-
-    return assignment_rec(lhs_src, new_rhs); // recrusive call
-  }
-  else if(
-    lhs.id() == ID_verilog_indexed_part_select_plus ||
-    lhs.id() == ID_verilog_indexed_part_select_minus)
-  {
-    // we flatten n-bit part select into n times extractbit
-    auto &part_select = to_verilog_indexed_part_select_plus_or_minus_expr(lhs);
-
-    const exprt &lhs_src = part_select.src();
-    const exprt &lhs_index = part_select.index();
-    const exprt &lhs_width = part_select.width();
-
-    auto index_opt = synthesis_constant(lhs_index);
-
-    if(!index_opt.has_value())
-    {
-      throw errort().with_location(lhs_index.source_location())
-        << "failed to convert part select index";
-    }
-
-    auto width_opt = synthesis_constant(lhs_width);
-
-    if(!width_opt.has_value())
-    {
-      throw errort().with_location(lhs_width.source_location())
-        << "failed to convert part select width";
-    }
-
-    mp_integer index = *index_opt, width = *width_opt;
-
-    // turn
-    //   a[i +: w]=e  (selects bits i to i+w-1)
-    //   a[i -: w]=e  (selects bits i-w+1 to i)
-    // into
-    //   a'==a WITH [lo:=e[0]] ... WITH [lo+w-1:=e[w-1]]
-
-    mp_integer lo, hi;
-
-    if(lhs.id() == ID_verilog_indexed_part_select_plus)
-    {
-      lo = index;
-      hi = index + width - 1;
-    }
-    else // ID_verilog_indexed_part_select_minus
-    {
-      lo = index - width + 1;
-      hi = index;
-    }
-
-    // start with 'a'
-    auto new_rhs = synth_expr(lhs_src, symbol_statet::FINAL);
-    const auto rhs_width = get_width(lhs_src.type());
-
-    // We drop bits that are out of bounds
-    auto from_in_range = std::max(mp_integer{0}, lo);
-    auto to_in_range = std::min(rhs_width - 1, hi);
-
-    // now add the indexes in the range
-    for(mp_integer i = from_in_range; i <= to_in_range; ++i)
-    {
-      exprt offset = from_integer(i - lo, integer_typet{});
-      exprt rhs_extractbit = extractbit_exprt{rhs, std::move(offset)};
-      exprt count = from_integer(i, integer_typet{});
-      new_rhs = with_exprt{
-        std::move(new_rhs), std::move(count), std::move(rhs_extractbit)};
-    }
-
-    return assignment_rec(lhs_src, new_rhs); // recursive call
-  }
-  else if(lhs.id() == ID_member)
-  {
-    const auto &member_expr = to_member_expr(lhs);
-    const exprt &lhs_compound = member_expr.struct_op();
-    auto component_name = member_expr.get_component_name();
-
-    if(
-      lhs_compound.type().id() == ID_struct ||
-      lhs_compound.type().id() == ID_union)
-    {
-      // turn
-      //   s.m=e
-      // into
-      //   s'==s WITH [m:=e]
-      auto synth_compound = synth_expr(lhs_compound, symbol_statet::FINAL);
-
-      with_exprt new_rhs{
-        synth_compound, member_designatort{component_name}, rhs};
-
-      // recursive call
-      return assignment_rec(lhs_compound, new_rhs); // recursive call
-    }
-    else
-    {
-      throw errort() << "unexpected member lhs: " << lhs_compound.type().id();
-    }
-  }
-  else if(lhs.id() == ID_typecast)
-  {
-    // The cast is assumed to be a reinterpret cast.
-    // (T)lhs = rhs
-    auto &typecast_expr = to_typecast_expr(lhs);
-
-    // note that the LHS is not yet fully synthesised; we will need to lower
-    // the typecast
-    auto new_rhs = typecast_exprt{rhs, typecast_expr.op().type()};
-
-    auto new_rhs_lowered = verilog_lowering_cast(new_rhs);
-
-    return assignment_rec(typecast_expr.op(), new_rhs_lowered);
-  }
-  else
-  {
-    throw errort() << "unexpected lhs: " << lhs.id();
-  }
-}
-
-/*******************************************************************\
-
 Function: verilog_synthesist::replace_by_wire
   Inputs:
 
@@ -947,42 +371,41 @@ Function: verilog_synthesist::replace_by_wire
 
 \*******************************************************************/
 
-void verilog_synthesist::replace_by_wire(
-  exprt &what,
-  const symbolt &base)
+void verilog_synthesist::replace_by_wire(exprt &what, const symbolt &base)
 {
   PRECONDITION(what.is_not_nil());
 
   symbolt new_symbol;
-  
+
   do
   {
-    unsigned c=temporary_counter++;
-    new_symbol.name=id2string(base.name)+"_aux"+std::to_string(c);
-    new_symbol.base_name=id2string(base.base_name)+"_aux"+std::to_string(c);
-  }
-  while(symbol_table.symbols.find(new_symbol.name)!=symbol_table.symbols.end());
+    unsigned c = temporary_counter++;
+    new_symbol.name = id2string(base.name) + "_aux" + std::to_string(c);
+    new_symbol.base_name =
+      id2string(base.base_name) + "_aux" + std::to_string(c);
+  } while(symbol_table.symbols.find(new_symbol.name) !=
+          symbol_table.symbols.end());
 
-  new_symbol.type=what.type();
-  new_symbol.module=base.module;
-  new_symbol.mode=base.mode;
-  new_symbol.location=base.location;
-  new_symbol.value=nil_exprt();
-  new_symbol.is_auxiliary=true;
+  new_symbol.type = what.type();
+  new_symbol.module = base.module;
+  new_symbol.mode = base.mode;
+  new_symbol.location = base.location;
+  new_symbol.value = nil_exprt();
+  new_symbol.is_auxiliary = true;
   new_symbol.pretty_name = strip_verilog_root_prefix(new_symbol.name);
 
   symbol_exprt symbol_expr(new_symbol.name, new_symbol.type);
 
-  assignmentt &assignment=assignments[new_symbol.name];
-  assignment.next.value=what;
+  assignmentt &assignment = assignments[new_symbol.name];
+  assignment.next.value = what;
   new_wires.insert(new_symbol.name);
-  
+
   if(symbol_table.add(new_symbol))
   {
     throw errort() << "failed to add replace_by_wire symbol";
   }
-    
-  what=symbol_expr;
+
+  what = symbol_expr;
 }
 
 /*******************************************************************\
@@ -1055,19 +478,19 @@ void verilog_synthesist::assignment_member_rec(
         << "failed to convert range";
     }
 
-    if(from>to)
+    if(from > to)
       std::swap(from, to);
 
     member.push_back(mp_integer());
 
     // now add the indexes in the range
-    for(mp_integer i=from; i<=to; ++i)
+    for(mp_integer i = from; i <= to; ++i)
     {
       // do the value
-      member.back()=i;
+      member.back() = i;
       assignment_member_rec(lhs_bv, member, data);
     }
-    
+
     member.pop_back();
   }
   else if(
@@ -1135,16 +558,15 @@ Function: verilog_synthesist::disjoint
 
 \*******************************************************************/
 
-bool verilog_synthesist::disjoint(
-  const membert &a,
-  const membert &b)
+bool verilog_synthesist::disjoint(const membert &a, const membert &b)
 {
-  membert::const_iterator a_it=a.begin();
-  membert::const_iterator b_it=b.begin();
-  
-  while(a_it!=a.end() && b_it!=b.end())
+  membert::const_iterator a_it = a.begin();
+  membert::const_iterator b_it = b.begin();
+
+  while(a_it != a.end() && b_it != b.end())
   {
-    if(*a_it!=*b_it) return true;
+    if(*a_it != *b_it)
+      return true;
     a_it++, b_it++;
   }
   
@@ -1271,401 +693,6 @@ const symbolt &verilog_synthesist::assignment_symbol(const exprt &lhs)
 
 /*******************************************************************\
 
-Function: verilog_synthesist::instantiate_port
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::instantiate_port(
-  const module_typet::portt &port,
-  const exprt &value,
-  const source_locationt &source_location,
-  transt &trans)
-{
-  symbol_exprt port_symbol{port.identifier(), port.type()};
-
-  // Much like
-  //   always @(*) port = value for an input, and
-  //   always @(*) value = port for an output.
-  // Note that the types need not match.
-  exprt lhs, rhs;
-
-  if(port.output())
-  {
-    lhs = value;
-    rhs = typecast_exprt::conditional_cast(port_symbol, value.type());
-  }
-  else
-  {
-    lhs = port_symbol;
-    rhs = typecast_exprt::conditional_cast(value, port_symbol.type());
-  }
-
-  verilog_forcet assignment{lhs, rhs};
-
-  assignment.add_source_location() = source_location;
-
-  verilog_event_guardt event_guard;
-  event_guard.add_source_location() = source_location;
-  event_guard.body() = assignment;
-
-  verilog_always_baset always(ID_verilog_always, event_guard, source_location);
-  synth_always_base(always);
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::instantiate_ports
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::instantiate_ports(
-  const irep_idt &instance,
-  const verilog_instt::instancet &inst,
-  const symbolt &symbol,
-  transt &trans)
-{
-  if(inst.connections().empty())
-    return;
-
-  auto &module_type = to_module_type(symbol.type);
-
-  // named port connection?
-
-  if(inst.named_port_connections())
-  {
-    const auto &ports = module_type.ports();
-    auto port_map = module_type.port_map();
-
-    // no requirement that all ports are connected
-    for(const auto &connection : inst.connections())
-    {
-      if(connection.id() == ID_verilog_wildcard_port_connection)
-        throw errort{}.with_location(connection.source_location())
-          << "no support for wildcard port connection";
-
-      auto &named_connection = to_verilog_named_port_connection(connection);
-      auto port_it =
-        port_map.find(to_symbol_expr(named_connection.port()).get_identifier());
-      CHECK_RETURN(port_it != port_map.end());
-      auto &port = port_it->second;
-      const exprt &value = named_connection.value();
-
-      if(value.is_not_nil())
-      {
-        instantiate_port(port, value, inst.source_location(), trans);
-      }
-    }
-
-    std::set<irep_idt> connected_ports;
-
-    for(const auto &connection : inst.connections())
-    {
-      auto &named_connection = to_verilog_named_port_connection(connection);
-      connected_ports.insert(
-        to_symbol_expr(named_connection.port()).get_identifier());
-    }
-
-    // unconnected inputs may be given a default value
-    for(auto &port : ports)
-      if(port.input())
-      {
-        auto identifier = port.identifier();
-        if(connected_ports.find(identifier) == connected_ports.end())
-        {
-          auto &port_symbol = ns.lookup(identifier);
-          if(port_symbol.value.is_not_nil())
-            instantiate_port(
-              port, port_symbol.value, inst.source_location(), trans);
-        }
-      }
-  }
-  else // just a list without names
-  {
-    const auto &ports = module_type.ports();
-
-    if(inst.connections().size() != ports.size())
-    {
-      throw errort().with_location(inst.source_location())
-        << "wrong number of ports: expected " << ports.size() << " but got "
-        << inst.connections().size();
-    }
-
-    auto p_it = ports.begin();
-
-    for(const auto &connection : inst.connections())
-    {
-      // 1800-2017 23.3.3.2 says
-      // "If left unconnected, the port shall have the default
-      // initial value corresponding to the data type."
-      // Simulators don't agree on this, and we hence leave the port
-      // unconstrained.
-      if(connection.is_not_nil())
-      {
-        instantiate_port(*p_it, connection, inst.source_location(), trans);
-      }
-
-      p_it++;
-    }
-  }
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::synth_module_instance
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::synth_module_instance(
-  const verilog_instt &statement,
-  transt &trans_dest)
-{
-  for(auto &instance : statement.instances())
-  {
-    const irep_idt &module_identifier = instance.module_identifier();
-
-    // must be in symbol_table
-    const symbolt &module_symbol = ns.lookup(module_identifier);
-
-    // get the transition relation of the instantiated module
-    auto trans_inst = verilog_synthesis(
-      symbol_table,
-      module_identifier,
-      standard,
-      ignore_initial,
-      initial_zero,
-      get_message_handler());
-
-    expand_module_instance(module_symbol, trans_inst, instance, trans_dest);
-  }
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::synth_module_instance_builtin
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::synth_module_instance_builtin(
-  const verilog_inst_builtint &module_item,
-  transt &trans)
-{
-  const irep_idt &module = module_item.module_base_name();
-
-  for(auto &instance : module_item.instances())
-  {
-    // check built-in ones
-    if(module==ID_bufif0 ||
-       module==ID_bufif1 ||
-       module==ID_notif0 ||
-       module==ID_notif1)
-    {
-      // add to general constraints
-
-      exprt constraint = instance;
-      constraint.id("verilog-primitive-module");
-      constraint.type()=bool_typet();
-      constraint.set(ID_module, module);
-    
-      assert(trans.operands().size()==3);
-      trans.invar().add_to_operands(std::move(constraint));
-    }
-    else if(module==ID_nmos ||
-            module==ID_pmos ||
-            module==ID_rnmos ||
-            module==ID_rpmos)
-    {
-      // add to general constraints
-
-      exprt constraint = instance;
-      constraint.id("verilog-primitive-module");
-      constraint.type()=bool_typet();
-      constraint.set(ID_module, module);
-    
-      assert(trans.operands().size()==3);
-      trans.invar().add_to_operands(std::move(constraint));
-    }
-    else if(
-      module == ID_and || module == ID_nand || module == ID_or ||
-      module == ID_nor || module == ID_xor || module == ID_xnor)
-    {
-      // 1800-2017 28.4 and, nand, nor, or, xor, and xnor gates
-      DATA_INVARIANT(
-        instance.connections().size() >= 2,
-        "binary primitive gates should have at least two connections");
-
-      // One output, one or more inputs.
-      auto &connections = instance.connections();
-      auto &output = connections[0];
-
-      irep_idt id = instance.type().id() == ID_bool
-                      ? module
-                      : irep_idt{"bit" + id2string(module)};
-
-      exprt::operandst operands;
-
-      // iterate over the module inputs
-      for(std::size_t i = 1; i < connections.size(); i++)
-      {
-        operands.push_back(connections[i]);
-      }
-
-      auto op = exprt{id, instance.type(), std::move(operands)};
-
-      equal_exprt constraint{output, op};
-      trans.invar().add_to_operands(std::move(constraint));
-    }
-    else if(module==ID_buf)
-    {
-      assert(instance.connections().size() >= 2);
-
-      for(unsigned i = 0; i < instance.connections().size() - 1; i++)
-      {
-        equal_exprt constraint{
-          instance.connections()[i], instance.connections().back()};
-
-        assert(trans.operands().size()==3);
-        trans.invar().add_to_operands(std::move(constraint));
-      }
-    }
-    else if(module==ID_not)
-    {
-      assert(instance.connections().size() >= 2);
-
-      // May have multiple outputs. The input is the last connection.
-      auto &input = instance.connections().back();
-      exprt rhs;
-
-      if(input.type().id() == ID_bool)
-        rhs = not_exprt{input};
-      else
-        rhs = bitnot_exprt{input};
-
-      rhs.add_source_location() = module_item.source_location();
-
-      for(std::size_t i = 0; i < instance.connections().size() - 1; i++)
-      {
-        auto &lhs = instance.connections()[i];
-        auto constraint = equal_exprt{lhs, rhs};
-        trans.invar().add_to_operands(std::move(constraint));
-      }
-    }
-    else if(module=="tranif0" ||
-            module=="tranif1" ||
-            module=="rtranif1" ||
-            module=="rtranif0")
-    {
-      // add to general constraints
-
-      exprt constraint = instance;
-      constraint.id("verilog-primitive-module");
-      constraint.type()=bool_typet();
-      constraint.set(ID_module, module);
-    
-      assert(trans.operands().size()==3);
-      trans.invar().add_to_operands(std::move(constraint));
-    }
-    else if(module=="tran"  ||
-            module=="rtran")
-    {
-      // add to general constraints
-
-      exprt constraint = instance;
-      constraint.id("verilog-primitive-module");
-      constraint.type()=bool_typet();
-      constraint.set(ID_module, module);
-    
-      assert(trans.operands().size()==3);
-      trans.invar().add_to_operands(std::move(constraint));
-    }
-    else
-    {
-      throw errort().with_location(module_item.source_location())
-        << "do not know how to synthesize " << module;
-    }
-  }
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::find_module_symbols
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-std::vector<irep_idt>
-verilog_synthesist::find_module_symbols(const symbolt &module_symbol) const
-{
-  std::vector<irep_idt> result;
-
-  auto lower = symbol_table.symbol_module_map.lower_bound(module_symbol.module);
-  auto upper = symbol_table.symbol_module_map.upper_bound(module_symbol.module);
-
-  for(auto it = lower; it != upper; it++)
-  {
-    result.push_back(it->second);
-  }
-
-  return result;
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::expand_module_instance
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::expand_module_instance(
-  const symbolt &module_symbol,
-  const transt &trans_inst,
-  const verilog_instt::instancet &instance,
-  transt &trans_dest)
-{
-  construct=constructt::OTHER;
-
-  // do the trans of the instantiated module
-  for(std::size_t i = 0; i < 3; i++)
-    trans_dest.operands()[i].add_to_operands(trans_inst.operands()[i]);
-
-  instantiate_ports(instance.base_name(), instance, module_symbol, trans_dest);
-}
-
-/*******************************************************************\
-
 Function: verilog_synthesist::synth_always_base
 
   Inputs:
@@ -1694,22 +721,22 @@ void verilog_synthesist::synth_always_base(
     DATA_INVARIANT(
       false, "unknown always construct: " + module_item.id_string());
 
-  event_guard=event_guardt::NONE;
+  event_guard = event_guardt::NONE;
 
   value_mapt always_value_map;
   DATA_INVARIANT(value_map == nullptr, "always/initial must not nest");
-  value_map=&always_value_map;
+  value_map = &always_value_map;
 
   synth_statement(module_item.statement());
 
-  for(const auto & it : value_map->final.changed)
+  for(const auto &it : value_map->final.changed)
   {
-    assignmentt &assignment=assignments[it];
-    assignment.next.value=value_map->final.symbol_map[it];
+    assignmentt &assignment = assignments[it];
+    assignment.next.value = value_map->final.symbol_map[it];
     assignment.next.move_assignments();
   }
 
-  value_map=NULL;
+  value_map = NULL;
 }
 
 /*******************************************************************\
@@ -1724,10 +751,9 @@ Function: verilog_synthesist::synth_initial
 
 \*******************************************************************/
 
-void verilog_synthesist::synth_initial(
-  const verilog_initialt &module_item)
+void verilog_synthesist::synth_initial(const verilog_initialt &module_item)
 {
-  if(module_item.operands().size()!=1)
+  if(module_item.operands().size() != 1)
   {
     throw errort().with_location(module_item.source_location())
       << "initial module item expected to have one operand";
@@ -1736,23 +762,23 @@ void verilog_synthesist::synth_initial(
   if(ignore_initial)
     return;
 
-  construct=constructt::INITIAL;
-  event_guard=event_guardt::NONE;
+  construct = constructt::INITIAL;
+  event_guard = event_guardt::NONE;
 
   value_mapt initial_value_map;
   DATA_INVARIANT(value_map == nullptr, "always/initial must not nest");
-  value_map=&initial_value_map;
+  value_map = &initial_value_map;
 
   synth_statement(module_item.statement());
-  
-  for(const auto & it : value_map->final.changed)
+
+  for(const auto &it : value_map->final.changed)
   {
-    assignmentt &assignment=assignments[it];
-    assignment.init.value=value_map->final.symbol_map[it];
-    assignment.init.move_assignments();    
+    assignmentt &assignment = assignments[it];
+    assignment.init.value = value_map->final.symbol_map[it];
+    assignment.init.move_assignments();
   }
 
-  value_map=NULL;
+  value_map = NULL;
 }
 
 /*******************************************************************\
@@ -1957,68 +983,6 @@ void verilog_synthesist::synth_function_or_task_decl(
 
 /*******************************************************************\
 
-Function: verilog_synthesist::synth_block
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::synth_block(const verilog_blockt &statement)
-{
-  for(auto &block_statement : statement.statements())
-    synth_statement(block_statement);
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::synth_break
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::synth_break(const verilog_breakt &statement)
-{
-  PRECONDITION(loop_frame.has_value());
-
-  loop_frame.value().break_statement_states.push_back(*value_map);
-
-  // set guard to false
-  value_map->guard.push_back(false_exprt{});
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::synth_continue
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::synth_continue(const verilog_continuet &statement)
-{
-  PRECONDITION(loop_frame.has_value());
-
-  loop_frame.value().continue_statement_states.push_back(*value_map);
-
-  // set guard to false
-  value_map->guard.push_back(false_exprt{});
-}
-
-/*******************************************************************\
-
 Function: verilog_synthesist::synth_continuous_assign
 
   Inputs:
@@ -2032,11 +996,11 @@ Function: verilog_synthesist::synth_continuous_assign
 void verilog_synthesist::synth_continuous_assign(
   const verilog_continuous_assignt &module_item)
 {
-  construct=constructt::OTHER;
+  construct = constructt::OTHER;
 
   forall_operands(it, module_item)
   {
-    if(it->id()!=ID_equal || it->operands().size()!=2)
+    if(it->id() != ID_equal || it->operands().size() != 2)
     {
       throw errort().with_location(it->source_location())
         << "unexpected continuous assignment";
@@ -2048,10 +1012,10 @@ void verilog_synthesist::synth_continuous_assign(
 
     assignment.lhs() = to_equal_expr(*it).lhs();
     assignment.rhs() = to_equal_expr(*it).rhs();
-    assignment.add_source_location()=module_item.source_location();    
-    
+    assignment.add_source_location() = module_item.source_location();
+
     verilog_event_guardt event_guard;
-    event_guard.add_source_location()=module_item.source_location();
+    event_guard.add_source_location() = module_item.source_location();
     event_guard.body() = assignment;
 
     verilog_always_baset always(
@@ -2089,11 +1053,9 @@ Function: verilog_synthesist::synth_force
 
 \*******************************************************************/
 
-void verilog_synthesist::synth_force_rec(
-  const exprt &lhs,
-  const exprt &rhs)
+void verilog_synthesist::synth_force_rec(const exprt &lhs, const exprt &rhs)
 {
-  if(lhs.id()==ID_concatenation)
+  if(lhs.id() == ID_concatenation)
   {
     // split it up
     mp_integer offset = 0;
@@ -2169,76 +1131,6 @@ void verilog_synthesist::synth_force_rec(
 
   equal_exprt equality{std::move(lhs_synth), std::move(rhs_synth)};
   invars.push_back(std::move(equality));
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::synth_assign
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::synth_assign(const verilog_assignt &statement)
-{
-  if(construct == constructt::OTHER)
-  {
-    throw errort().with_location(statement.source_location())
-      << "unexpected assignment statement";
-  }
-
-  exprt lhs = statement.lhs();
-  lhs = synth_lhs_expr(lhs);
-
-  exprt rhs = statement.rhs();
-  rhs = synth_expr(rhs, symbol_statet::CURRENT);
-
-  irep_idt compound_id = irep_idt{};
-
-  if(statement.id() == ID_verilog_blocking_assign_plus)
-    compound_id = ID_plus;
-  else if(statement.id() == ID_verilog_blocking_assign_minus)
-    compound_id = ID_minus;
-  else if(statement.id() == ID_verilog_blocking_assign_mult)
-    compound_id = ID_mult;
-  else if(statement.id() == ID_verilog_blocking_assign_div)
-    compound_id = ID_div;
-  else if(statement.id() == ID_verilog_blocking_assign_mod)
-    compound_id = ID_mod;
-  else if(statement.id() == ID_verilog_blocking_assign_bitand)
-    compound_id = ID_bitand;
-  else if(statement.id() == ID_verilog_blocking_assign_bitor)
-    compound_id = ID_bitor;
-  else if(statement.id() == ID_verilog_blocking_assign_bitxor)
-    compound_id = ID_bitxor;
-  else if(statement.id() == ID_verilog_blocking_assign_lshr)
-    compound_id = ID_lshr;
-  else if(statement.id() == ID_verilog_blocking_assign_lshl)
-    compound_id = ID_shl;
-  else if(statement.id() == ID_verilog_blocking_assign_ashr)
-    compound_id = ID_ashr;
-  else if(statement.id() == ID_verilog_blocking_assign_ashl)
-    compound_id = ID_shl;
-
-  if(compound_id != irep_idt{})
-  {
-    auto lhs_synth = synth_expr(lhs, symbol_statet::CURRENT);
-    rhs = binary_exprt{std::move(lhs_synth), compound_id, rhs, rhs.type()};
-  }
-
-  // Can the rhs be simplified to a constant, for propagation?
-  auto rhs_simplified = simplify_expr(rhs, ns);
-
-  if(rhs_simplified.is_constant())
-    rhs = rhs_simplified;
-
-  bool blocking = statement.id() != ID_verilog_non_blocking_assign;
-
-  assignment_rec(lhs, rhs, blocking);
 }
 
 /*******************************************************************\
@@ -2520,915 +1412,6 @@ void verilog_synthesist::synth_assume(
 
   // add it as an invariant
   invars.push_back(condition);
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::case_comparison
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-exprt verilog_synthesist::case_comparison(
-  const exprt &case_operand,
-  const exprt &pattern)
-{
-  // the pattern has the max type, not the case operand
-  const typet &pattern_type=pattern.type();
-
-  // we need to take case of ?, x, z in the pattern
-  if(is_aval_bval(pattern_type))
-  {
-    // We are using masking based on the pattern.
-    // The aval is the comparison value, and the
-    // negation of bval is the mask.
-    auto pattern_aval = ::aval(pattern);
-    auto pattern_bval = ::bval(pattern);
-    auto mask_expr = bitnot_exprt{pattern_bval};
-
-    auto case_operand_casted = typecast_exprt{
-      typecast_exprt::conditional_cast(
-        case_operand, aval_bval_underlying(pattern_type)),
-      mask_expr.type()};
-
-    return equal_exprt{
-      bitand_exprt{case_operand_casted, mask_expr},
-      bitand_exprt{pattern_aval, mask_expr}};
-  }
-
-  // 2-valued comparison
-  exprt case_operand_casted =
-    typecast_exprt::conditional_cast(case_operand, pattern_type);
-
-  return equal_exprt(case_operand_casted, pattern);
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::synth_case_values
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-exprt verilog_synthesist::synth_case_values(
-  const exprt &values,
-  const exprt &case_operand)
-{
-  if(values.id()==ID_default)
-    return true_exprt();
-
-  exprt::operandst op;
-  
-  op.reserve(values.operands().size());
-
-  forall_operands(it, values)
-  {
-    auto pattern = synth_expr(*it, symbol_statet::CURRENT);
-    op.push_back(case_comparison(case_operand, pattern));
-  }
-
-  return disjunction(op);
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::synth_case
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::synth_case(
-  const verilog_case_statement_baset &statement)
-{
-  if(statement.operands().size()<1)
-  {
-    throw errort().with_location(statement.source_location())
-      << "case statement expected to have at least one operand";
-  }
-
-  // do the argument of the case
-  auto case_operand =
-    synth_expr(statement.case_operand(), symbol_statet::CURRENT);
-
-  // we convert the rest to if-then-else
-  exprt start;
-  exprt *last_if=&start;
-  std::optional<verilog_statementt> default_case;
-
-  for(unsigned i=1; i<statement.operands().size(); i++)
-  {
-    const exprt &e=statement.operands()[i];
-
-    if(e.id()!=ID_case_item)
-    {
-      throw errort() << "expected case_item";
-    }
-
-    auto &case_item = to_verilog_case_item(e);
-
-    // default?
-    if(case_item.is_default())
-      default_case = case_item.case_statement();
-    else
-    {
-      exprt if_statement(ID_if);
-      if_statement.reserve_operands(3);
-      if_statement.add_to_operands(
-        synth_case_values(case_item.case_value(), case_operand));
-      if_statement.add_to_operands(case_item.case_statement());
-
-      last_if->add_to_operands(std::move(if_statement));
-      last_if = &last_if->operands().back();
-    }
-  }
-
-  if(default_case.has_value())
-  {
-    // Attach the 'default' to the final 'if' as 'else'
-    last_if->add_to_operands(std::move(default_case.value()));
-  }
-
-  // synthesize the new if-then-else
-
-  if(!start.operands().empty())
-    synth_statement(
-      static_cast<verilog_statementt &>(to_multi_ary_expr(start).op0()));
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::synth_if
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::synth_if(
-  const verilog_ift &statement)
-{
-  auto if_cond = synth_expr(statement.cond(), symbol_statet::CURRENT);
-
-  if(if_cond.is_true())
-  {
-    synth_statement(statement.then_case());
-    return;
-  }
-
-  // save current value_map pointer
-  value_mapt *old_map=value_map;
-
-  // produce new value maps
-  value_mapt true_map(*value_map), false_map(*value_map);
-
-  true_map.clear_changed();
-  true_map.guard.push_back(if_cond);
-
-  false_map.clear_changed();
-  false_map.guard.push_back(not_exprt{if_cond});
-
-  // 'then' case
-  {
-    value_map=&true_map;
-    synth_statement(statement.then_case());
-  }
-
-  // 'else' case
-  if(statement.has_else_case())
-  {
-    value_map=&false_map;
-    synth_statement(statement.else_case());
-  }
-
-  // restore and merge
-  value_map=old_map;
-
-  // merge current map
-  merge(
-    if_cond, true_map.current, false_map.current, false, value_map->current);
-
-  // merge final map
-  merge(if_cond, true_map.final, false_map.final, true, value_map->final);
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::merge
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::merge(
-  const exprt &guard,
-  const value_mapt::mapt &true_map,
-  const value_mapt::mapt &false_map,
-  bool use_previous_assignments,
-  value_mapt::mapt &dest)
-{
-  // get true_map.changed + false_map.changed
-
-  std::set<irep_idt> changed;
-
-  changed.insert(true_map.changed.begin(), true_map.changed.end());
-  changed.insert(false_map.changed.begin(), false_map.changed.end());
-
-  for(const auto & it : changed)
-  {
-    const symbolt &symbol=ns.lookup(it);
-
-    exprt true_value=current_value(true_map, symbol, use_previous_assignments);
-    exprt false_value=current_value(false_map, symbol, use_previous_assignments);
-    
-    // this is a phi-node equivalent
-    if_exprt value{guard, true_value, false_value, symbol.type};
-
-    dest.symbol_map[symbol.name].swap(value);
-    dest.changed.insert(symbol.name);
-  }
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::synth_event_guard
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::synth_event_guard(
-  const verilog_event_guardt &statement)
-{
-  if(statement.operands().size()!=2)
-  {
-    throw errort().with_location(statement.source_location())
-      << "event_guard expected to have two operands";
-  }
-
-  if(event_guard!=event_guardt::NONE)
-  {
-    throw errort().with_location(statement.source_location())
-      << "event guard already specified";
-  }
-
-  const exprt &event_guard_expr=statement.guard();
-
-  bool edge=false;
-
-  // these guards are ORed
-  exprt::operandst guards; 
-
-  forall_operands(it, event_guard_expr)
-    if(it->id()==ID_posedge || it->id()==ID_negedge)
-    {
-      edge=true;
-
-      if(it->operands().size()!=1)
-      {
-        throw errort().with_location(it->source_location())
-          << "pos/negedge expected to have one operand";
-      }
-
-      irep_idt identifier="conf::clock_enable_mode";
-
-      // check symbol_table for clock guard
-
-      if(symbol_table.symbols.find(identifier)!=symbol_table.symbols.end())
-      {
-        // found! we make it a guard
-
-        auto &op = to_unary_expr(*it).op();
-
-        if(op.type().id() == ID_bool)
-          guards.push_back(op);
-        else
-        {
-          // get LSB
-          guards.push_back(extractbit_exprt{op, integer_typet{}.zero_expr()});
-        }
-
-        throw errort() << "Notice: using clock guard " << identifier;
-      }
-    }
-
-  event_guard=edge?event_guardt::CLOCK:event_guardt::COMBINATIONAL;
-
-  if(guards.empty())
-    synth_statement(statement.body());
-  else
-  {
-    // new guards!
-    exprt guard_expr=disjunction(guards);
-
-    value_mapt *old_map=value_map;
-    value_mapt true_map(*value_map), false_map(*value_map);
-    true_map.clear_changed();
-    false_map.clear_changed();
-
-    value_map=&true_map;
-    synth_statement(statement.body());
-
-    value_map=old_map;
-    merge(guard_expr, true_map.current, false_map.current, false, value_map->current);
-    merge(guard_expr, true_map.final, false_map.final, true, value_map->final);
-  }
-
-  event_guard=event_guardt::NONE;
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::synth_delay
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::synth_delay(
-  const verilog_delayt &statement)
-{
-  if(statement.operands().size()!=2)
-  {
-    throw errort().with_location(statement.source_location())
-      << "delay expected to have two operands";
-  }
-
-  synth_statement(statement.body());
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::synth_for
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::synth_for(const verilog_fort &statement)
-{
-  if(statement.operands().size()!=4)
-  {
-    throw errort().with_location(statement.source_location())
-      << "for expected to have four operands";
-  }
-
-  auto old_loop_frame = std::move(loop_frame);
-  loop_frame = loop_framet{};
-
-  for(auto &init : statement.initialization())
-  {
-    // either an assignment or a declaration with assignment
-    if(
-      init.id() == ID_verilog_blocking_assign ||
-      init.id() == ID_verilog_non_blocking_assign)
-    {
-      synth_statement(init);
-    }
-    else if(init.id() == ID_decl)
-    {
-      // turn into a blocking assignment
-      auto &decl = to_verilog_decl(init);
-      for(auto &declarator : decl.declarators())
-      {
-        DATA_INVARIANT(
-          declarator.value().is_not_nil(),
-          "for-init declarator must have value");
-        auto assignment = verilog_blocking_assignt{
-          declarator.symbol_expr(), declarator.value()};
-        synth_assign(assignment);
-      }
-    }
-    else
-    {
-      DATA_INVARIANT_WITH_DIAGNOSTICS(
-        false, "unexpected initialization in for loop", init.pretty());
-    }
-  }
-
-  while(true)
-  {
-    loop_frame.value().continue_statement_states.clear();
-
-    DATA_INVARIANT(
-      statement.condition().type().id() == ID_bool,
-      "for condition must be Boolean");
-
-    auto guard_value_opt = synthesis_constant(statement.condition());
-
-    if(!guard_value_opt.has_value())
-    {
-      throw errort().with_location(statement.condition().source_location())
-        << "synthesis failed to evaluate loop guard: `"
-        << to_string(statement.condition()) << '\'';
-    }
-
-    if(guard_value_opt.value() == 0)
-      break;
-
-    // execute the body
-    synth_statement(statement.body());
-
-    // merge in edges from 'continue' statements, if any
-    for(auto &state : loop_frame.value().continue_statement_states)
-    {
-      auto guard_expr = conjunction(state.guard);
-      merge(
-        guard_expr,
-        state.current,
-        value_map->current,
-        false,
-        value_map->current);
-      merge(guard_expr, state.final, value_map->final, true, value_map->final);
-    }
-
-    // execute the step statement
-    synth_statement(statement.inc_statement());
-  }
-
-  // Merge in edges from 'break' statements, if any. These come
-  // in program order, hence process in reverse order.
-  auto &break_states = loop_frame.value().break_statement_states;
-
-  for(auto state_it = break_states.rbegin(); state_it != break_states.rend();
-      ++state_it)
-  {
-    auto &state = *state_it;
-    auto guard_expr = conjunction(state.guard);
-    merge(
-      guard_expr, state.current, value_map->current, false, value_map->current);
-    merge(guard_expr, state.final, value_map->final, true, value_map->final);
-  }
-
-  loop_frame = std::move(old_loop_frame);
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::synth_prepostincdec
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::synth_prepostincdec(const verilog_statementt &statement)
-{
-  if(statement.operands().size()!=1)
-  {
-    throw errort().with_location(statement.source_location())
-      << statement.id() << " expected to have one operand";
-  }
-
-  const auto &op = to_unary_expr(statement).op();
-
-  // stupid implementation
-  exprt one = from_integer(1, op.type());
-
-  bool increment =
-    statement.id() == ID_preincrement || statement.id() == ID_postincrement;
-
-  exprt rhs;
-
-  if(increment)
-    rhs = plus_exprt(op, one);
-  else
-    rhs = minus_exprt(op, one);
-
-  verilog_blocking_assignt assignment{op, rhs};
-  assignment.add_source_location()=statement.source_location();
-
-  synth_statement(assignment);  
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::synth_while
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::synth_while(
-  const verilog_whilet &statement)
-{
-  if(statement.operands().size()!=2)
-  {
-    throw errort().with_location(statement.source_location())
-      << "while expected to have two operands";
-  }
-
-  auto old_loop_frame = std::move(loop_frame);
-  loop_frame = loop_framet{};
-
-  while(true)
-  {
-    loop_frame.value().continue_statement_states.clear();
-
-    exprt tmp_guard=statement.condition();
-    tmp_guard = typecast_exprt{tmp_guard, bool_typet{}};
-    tmp_guard = synth_expr(tmp_guard, symbol_statet::CURRENT);
-    simplify(tmp_guard, ns);
- 
-    if(!tmp_guard.is_constant())
-    {
-      throw errort().with_location(statement.body().source_location())
-        << "synthesis failed to evaluate loop guard: `" << to_string(tmp_guard)
-        << '\'';
-    }
-
-    if(tmp_guard.is_false()) break;
-
-    // execute the body
-    synth_statement(statement.body());
-
-    // merge in edges from 'continue' statements, if any
-    for(auto &state : loop_frame.value().continue_statement_states)
-    {
-      auto guard_expr = conjunction(state.guard);
-      merge(
-        guard_expr,
-        state.current,
-        value_map->current,
-        false,
-        value_map->current);
-      merge(guard_expr, state.final, value_map->final, true, value_map->final);
-    }
-  }
-
-  // Merge in edges from 'break' statements, if any. These come
-  // in program order, hence process in reverse order.
-  auto &break_states = loop_frame.value().break_statement_states;
-
-  for(auto state_it = break_states.rbegin(); state_it != break_states.rend();
-      ++state_it)
-  {
-    auto &state = *state_it;
-    auto guard_expr = conjunction(state.guard);
-    merge(
-      guard_expr, state.current, value_map->current, false, value_map->current);
-    merge(guard_expr, state.final, value_map->final, true, value_map->final);
-  }
-
-  loop_frame = std::move(old_loop_frame);
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::synth_repeat
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::synth_repeat(
-  const verilog_repeatt &statement)
-{
-  if(statement.operands().size()!=2)
-  {
-    throw errort().with_location(statement.source_location())
-      << "repeat expected to have two operands";
-  }
-
-  throw errort().with_location(statement.source_location())
-    << "cannot synthesize `repeat'";
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::synth_return
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::synth_return(const verilog_returnt &statement)
-{
-  // There has to be a tf frame
-  PRECONDITION(tf_frame.has_value());
-
-  auto &frame = tf_frame.value();
-
-  // return value?
-  if(statement.has_value())
-  {
-    // assign to the symbol with the function's name
-    DATA_INVARIANT(
-      frame.return_value.has_value(),
-      "return with value requires return value symbol");
-
-    verilog_assignt assignment{
-      ID_verilog_blocking_assign,
-      frame.return_value.value(),
-      statement.value()};
-
-    synth_assign(assignment);
-  }
-
-  frame.return_statement_states.push_back(*value_map);
-
-  // set guard to false
-  value_map->guard.push_back(false_exprt{});
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::synth_forever
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::synth_forever(
-  const verilog_forevert &statement)
-{
-  if(statement.operands().size()!=1)
-  {
-    throw errort().with_location(statement.source_location())
-      << "forever expected to have one operand";
-  }
-
-  throw errort().with_location(statement.source_location())
-    << "cannot synthesize `forever'";
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::synth_function_call_or_task_enable
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::synth_function_call_or_task_enable(
-  const verilog_function_callt &statement)
-{
-  if(statement.is_system_function_call())
-  {
-    // ignore system functions
-  }
-  else
-  {
-    // this is essentially inlined
-    const symbol_exprt &function = to_symbol_expr(statement.function());
-    const symbolt &symbol = ns.lookup(function);
-
-    if(symbol.type.id()!=ID_code)
-    {
-      throw errort().with_location(statement.source_location())
-        << "expected function or task as first operand";
-    }
-    
-    const code_typet &code_type=to_code_type(symbol.type);
-
-    // preserve the previous call frame, if any
-    auto old_tf_frame = tf_frame;
-
-    // remember the guard
-    auto entry_guard = value_map->guard;
-
-    // create a fresh call frame
-    tf_frame = tf_framet{};
-
-    const code_typet::parameterst &parameters=
-      code_type.parameters();
-
-    const exprt::operandst &actuals = statement.arguments();
-
-    if(parameters.size()!=actuals.size())
-    {
-      throw errort().with_location(statement.source_location())
-        << "wrong number of arguments";
-    }
-    
-    // do assignments to input parameters
-    for(unsigned i=0; i<parameters.size(); i++)
-    {
-      const symbolt &a_symbol=ns.lookup(parameters[i].get_identifier());
-      if(parameters[i].get_bool(ID_input))
-      {
-        verilog_blocking_assignt assignment{a_symbol.symbol_expr(), actuals[i]};
-        assignment.add_source_location()=statement.source_location();
-        synth_statement(assignment);
-      }
-    }
-
-    for(auto &statement : symbol.value.operands())
-      synth_statement(to_verilog_statement(statement));
-
-    // merge in edges from 'return' statements, if any
-    for(auto &state : tf_frame.value().return_statement_states)
-    {
-      auto guard_expr = conjunction(state.guard);
-      merge(
-        guard_expr,
-        state.current,
-        value_map->current,
-        false,
-        value_map->current);
-      merge(guard_expr, state.final, value_map->final, true, value_map->final);
-    }
-
-    // restore the previous task/function frame
-    tf_frame = old_tf_frame;
-    value_map->guard = entry_guard;
-
-    // do assignments to output parameters
-    for(unsigned i=0; i<parameters.size(); i++)
-    {
-      const symbolt &a_symbol=ns.lookup(parameters[i].get_identifier());
-      if(parameters[i].get_bool(ID_output))
-      {
-        verilog_blocking_assignt assignment{actuals[i], a_symbol.symbol_expr()};
-        assignment.add_source_location()=statement.source_location();
-        synth_statement(assignment);
-      }
-    }
-  }
-}
-
-/*******************************************************************\
-
-Function: verilog_synthesist::synth_statement
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-void verilog_synthesist::synth_statement(
-  const verilog_statementt &statement)
-{
-  if(statement.id()==ID_block)
-    synth_block(to_verilog_block(statement));
-  else if(statement.id() == ID_break)
-    synth_break(to_verilog_break(statement));
-  else if(
-    statement.id() == ID_verilog_case || statement.id() == ID_verilog_casex ||
-    statement.id() == ID_verilog_casez)
-  {
-    synth_case(to_verilog_case_statement_base(statement));
-  }
-  else if(
-    statement.id() == ID_verilog_blocking_assign ||
-    statement.id() == ID_verilog_blocking_assign_plus ||
-    statement.id() == ID_verilog_blocking_assign_minus ||
-    statement.id() == ID_verilog_blocking_assign_mult ||
-    statement.id() == ID_verilog_blocking_assign_div ||
-    statement.id() == ID_verilog_blocking_assign_mod ||
-    statement.id() == ID_verilog_blocking_assign_bitand ||
-    statement.id() == ID_verilog_blocking_assign_bitor ||
-    statement.id() == ID_verilog_blocking_assign_bitxor ||
-    statement.id() == ID_verilog_blocking_assign_lshr ||
-    statement.id() == ID_verilog_blocking_assign_lshl ||
-    statement.id() == ID_verilog_blocking_assign_ashr ||
-    statement.id() == ID_verilog_blocking_assign_ashl)
-  {
-    synth_assign(to_verilog_assign(statement));
-  }
-  else if(statement.id() == ID_continue)
-    synth_continue(to_verilog_continue(statement));
-  else if(statement.id() == ID_procedural_continuous_assign)
-  {
-    throw errort().with_location(statement.source_location())
-      << "synthesis of procedural continuous assignment not supported";
-  }
-  else if(
-    statement.id() == ID_verilog_immediate_assert ||
-    statement.id() == ID_verilog_assert_property ||
-    statement.id() == ID_verilog_smv_assert ||
-    statement.id() == ID_verilog_immediate_cover ||
-    statement.id() == ID_verilog_cover_property ||
-    statement.id() == ID_verilog_cover_sequence)
-  {
-    synth_assert_assume_cover(
-      to_verilog_assert_assume_cover_statement(statement));
-  }
-  else if(
-    statement.id() == ID_verilog_immediate_assume ||
-    statement.id() == ID_verilog_assume_property ||
-    statement.id() == ID_verilog_restrict_property ||
-    statement.id() == ID_verilog_smv_assume)
-  {
-    synth_assert_assume_cover(
-      to_verilog_assert_assume_cover_statement(statement));
-  }
-  else if(statement.id() == ID_verilog_expect_property)
-  {
-    throw errort().with_location(statement.source_location())
-      << "synthesis of expect property not supported";
-  }
-  else if(statement.id() == ID_verilog_non_blocking_assign)
-    synth_assign(to_verilog_assign(statement));
-  else if(statement.id()==ID_force)
-    synth_force(to_verilog_force(statement));
-  else if(statement.id()==ID_if)
-    synth_if(to_verilog_if(statement));
-  else if(statement.id()==ID_event_guard)
-    synth_event_guard(to_verilog_event_guard(statement));
-  else if(statement.id()==ID_delay)
-    synth_delay(to_verilog_delay(statement));
-  else if(statement.id()==ID_for)
-    synth_for(to_verilog_for(statement));
-  else if(statement.id()==ID_while)
-    synth_while(to_verilog_while(statement));
-  else if(statement.id()==ID_repeat)
-    synth_repeat(to_verilog_repeat(statement));
-  else if(statement.id() == ID_return)
-    synth_return(to_verilog_return(statement));
-  else if(statement.id()==ID_forever)
-    synth_forever(to_verilog_forever(statement));
-  else if(statement.id()==ID_function_call)
-    synth_function_call_or_task_enable(to_verilog_function_call(statement));
-  else if(statement.id()==ID_preincrement ||
-          statement.id()==ID_predecrement ||
-          statement.id()==ID_postincrement ||
-          statement.id()==ID_postdecrement)
-    synth_prepostincdec(statement);
-  else if(statement.id()==ID_decl)
-  {
-    auto decl_class = to_verilog_decl(statement).get_class();
-
-    if(decl_class == ID_function || decl_class == ID_task)
-    {
-    }
-    else
-      synth_decl(to_verilog_decl(statement));
-  }
-  else if(statement.id()==ID_skip)
-  {
-    // do nothing
-  }
-  else if(statement.id() == ID_verilog_label_statement)
-    synth_statement(to_verilog_label_statement(statement).statement());
-  else if(statement.id() == ID_verilog_event_trigger)
-  {
-    // not synthesized
-  }
-  else
-  {
-    throw errort().with_location(statement.source_location())
-      << "unexpected statement during synthesis: " << statement.id();
-  }
 }
 
 /*******************************************************************\
